requirements.txt
Django==3.1.7
djangorestframework==3.12.2
psycopg2-binary==2.8.6
django-cors-headers==3.7.0
django-extensions==3.1.1
Pillow==8.1.2


